requirements.txt
numpy
pyspi
pyarrow
pandas
dill
scikit-learn
tqdm
pyyaml